feat: remove unused codeblock and duplicate variables
