+++ kubeflow/trainer/rhai/transformers.py
@@ -951,6 +951,7 @@ def get_trainer_cr_from_transformers_trainer(
 
 def _build_checkpoint_code(trainer: TransformersTrainer) -> str:
     """Generate checkpoint injection code for the trainer."""
+    from kubeflow.trainer.rhai.constants import S3_URI_SCHEME
 
     # Only inject if JIT or periodic checkpoint is enabled
     if not trainer.enable_jit_checkpoint and not trainer.periodic_checkpoint_config:
@@ -975,9 +976,16 @@ def _build_checkpoint_code(trainer: TransformersTrainer) -> str:
 
     resolved_output_dir, _ = parse_output_dir_uri(trainer.output_dir)
 
+    # Check if using S3 storage
+    storage_uri = None
+    if trainer.output_dir and trainer.output_dir.startswith(S3_URI_SCHEME):
+        storage_uri = trainer.output_dir
+
+
     # Generate checkpoint injection code
     return get_jit_checkpoint_injection_code(
         output_dir=resolved_output_dir,
+        storage_uri=storage_uri,
         periodic_checkpoint_config=periodic_config_dict,
         enable_jit_checkpoint=trainer.enable_jit_checkpoint,
     )
@@ -985,6 +993,7 @@ def _build_checkpoint_code(trainer: TransformersTrainer) -> str:
 
 def get_jit_checkpoint_injection_code(
     output_dir: Optional[str] = None,
+    storage_uri: Optional[str] = None,
     periodic_checkpoint_config: Optional[dict] = None,
     enable_jit_checkpoint: bool = False,
 ) -> str:
@@ -996,6 +1005,9 @@ def get_jit_checkpoint_injection_code(
 
     if output_dir:
         config_dict["output_dir"] = output_dir
+
+    if storage_uri:
+        config_dict["storage_uri"] = storage_uri
 
     if periodic_checkpoint_config:
         if "save_strategy" in periodic_checkpoint_config:
