+++ kubeflow/trainer/rhai/utils.py
@@ -10,9 +10,7 @@ from kubeflow.trainer.rhai import (
     transformers,
 )
 from kubeflow.trainer.rhai.constants import (
-    CHECKPOINT_EPHEMERAL_MOUNT_PATH,
     CHECKPOINT_EPHEMERAL_STORAGE_CLASS,
-    CHECKPOINT_EPHEMERAL_VOLUME_NAME,
     CHECKPOINT_EPHEMERAL_VOLUME_SIZE,
     CHECKPOINT_MOUNT_PATH,
     CHECKPOINT_VOLUME_NAME,
@@ -21,6 +19,7 @@ from kubeflow.trainer.rhai.constants import (
     S3_URI_SCHEME,
 )
 from kubeflow.trainer.types import types
+from kubernetes import client
 
 logger = logging.getLogger(__name__)
 
@@ -138,31 +137,35 @@ def parse_output_dir_uri(output_dir: Optional[str]) -> tuple[Optional[str], Opti
     if output_dir.startswith(S3_URI_SCHEME):
         # Build ephemeral volume spec for S3 checkpoint staging
         # This volume is used as temporary local storage before uploading to S3
+        volume_resources = {
+            "requests": {
+                "storage": models.IoK8sApimachineryPkgApiResourceQuantity(
+                    CHECKPOINT_EPHEMERAL_VOLUME_SIZE
+                ),
+            }
+        }
+
         volume_spec = {
-            "name": CHECKPOINT_EPHEMERAL_VOLUME_NAME,
+            "accessModes": ["ReadWriteOnce"],
+            "storageClassName": CHECKPOINT_EPHEMERAL_STORAGE_CLASS,
+            "resources": volume_resources
+        }
+
+        volume = {
+            "name": CHECKPOINT_VOLUME_NAME,
             "ephemeral": {
                 "volumeClaimTemplate": {
-                    "spec": {
-                        "accessModes": ["ReadWriteOnce"],
-                        "storageClassName": CHECKPOINT_EPHEMERAL_STORAGE_CLASS,
-                        "resources": {
-                            "requests": {
-                                "storage": models.IoK8sApimachineryPkgApiResourceQuantity(
-                                    CHECKPOINT_EPHEMERAL_VOLUME_SIZE
-                                ),
-                            }
-                        },
-                    }
+                "spec": volume_spec
                 }
             },
         }
         volume_mount_spec = {
-            "name": CHECKPOINT_EPHEMERAL_VOLUME_NAME,
-            "mountPath": CHECKPOINT_EPHEMERAL_MOUNT_PATH,
+            "name": CHECKPOINT_VOLUME_NAME,
+            "mountPath": CHECKPOINT_MOUNT_PATH,
         }
 
         # Return local staging path (training writes here, then uploads to S3)
-        return CHECKPOINT_EPHEMERAL_MOUNT_PATH, {"volume": volume_spec, "volumeMount": volume_mount_spec}
+        return CHECKPOINT_MOUNT_PATH, {"volume": volume, "volumeMount": volume_mount_spec}
 
     return output_dir, None
 
@@ -291,3 +294,30 @@ def get_s3_credential_env_vars(
     return env_vars
 
 
+def validate_secret_exists(
+    core_api: "client.CoreV1Api",
+    secret_name: str,
+    namespace: str,
+) -> None:
+    """Validate that a Kubernetes secret exists.
+
+    Args:
+        core_api: Kubernetes CoreV1Api client.
+        secret_name: Name of the secret to check.
+        namespace: Namespace to check in.
+
+    Raises:
+        ValueError: If secret does not exist.
+    """
+    from kubernetes.client.rest import ApiException
+
+    try:
+        core_api.read_namespaced_secret(name=secret_name, namespace=namespace)
+    except ApiException as e:
+        if e.status == 404:
+            raise ValueError(
+                f"Secret '{secret_name}' not found in namespace '{namespace}'. "
+                "Please create the Data Connection secret or verify the "
+                "data_connection_name is correct."
+            ) from e
+        raise
